ME-721: Migrated get_configuration.
